SBERDOMA-47 DDD refactor
