+++ apps/condo/domains/billing/schema/BillingAccount.js
@@ -8,10 +8,10 @@ const { GQLListSchema } = require('@core/keystone/schema')
 const { historical, versioned, uuided, tracked, softDeleted } = require('@core/keystone/plugins')
 const { SENDER_FIELD, DV_FIELD } = require('@condo/domains/common/schema/fields')
 const access = require('@condo/domains/billing/access/BillingAccount')
-const { DV_UNKNOWN_VERSION_ERROR } = require('../../common/constants/errors')
-const { hasRequestAndDbFields } = require('../../../utils/validation.utils')
-const { JSON_EXPECT_OBJECT_ERROR } = require('../../common/constants/errors')
-const { JSON_UNKNOWN_VERSION_ERROR } = require('../../common/constants/errors')
+const { DV_UNKNOWN_VERSION_ERROR } = require('@condo/domains/common/constants/errors')
+const { hasRequestAndDbFields } = require('@condo/domains/common/utils/validation.utils')
+const { JSON_EXPECT_OBJECT_ERROR } = require('@condo/domains/common/constants/errors')
+const { JSON_UNKNOWN_VERSION_ERROR } = require('@condo/domains/common/constants/errors')
 
 const { INTEGRATION_CONTEXT_FIELD, IMPORT_ID_FIELD, BILLING_PROPERTY_FIELD, RAW_DATA_FIELD } = require('./fields')
