+++ apps/condo/domains/billing/schema/BillingIntegrationOrganizationContext.js
@@ -8,7 +8,7 @@ const { GQLListSchema } = require('@core/keystone/schema')
 const { historical, versioned, uuided, tracked, softDeleted } = require('@core/keystone/plugins')
 const { SENDER_FIELD, DV_FIELD } = require('@condo/domains/common/schema/fields')
 const access = require('@condo/domains/billing/access/BillingIntegrationOrganizationContext')
-const { hasValidJsonStructure } = require('../../../utils/validation.utils')
+const { hasValidJsonStructure } = require('@condo/domains/common/utils/validation.utils')
 
 // TODO(pahaz): move if after Organization refactoring
 const { ORGANIZATION_OWNED_FIELD } = require('../../../schema/_common')
