+++ apps/condo/domains/common/components/containers/BaseLayout/BaseLayout.tsx
@@ -14,7 +14,7 @@ import classnames from 'classnames'
 import 'antd/dist/antd.less'
 import { TopMenuItems } from './components/TopMenuItems'
 import { useIntl } from '@core/next/intl'
-import { useAntdMediaQuery } from '../../../../../utils/mediaQuery.utils'
+import { useAntdMediaQuery } from '../../../utils/mediaQuery.utils'
 import { layoutCss, pageContentCss, pageHeaderCss, pageWrapperCss, subLayoutCss, topMenuCss } from './components/styles'
 
 const LayoutContext = createContext({})
