+++ apps/condo/domains/common/components/containers/FormTableExcelImport.tsx
@@ -4,7 +4,7 @@ import React, { useEffect, useState } from 'react'
 import XLSX from 'xlsx'
 import { Button, Col, Form, Input, Progress, Row, Select, Table, Tooltip, Typography, Upload } from 'antd'
 import { InboxOutlined } from '@ant-design/icons'
-import { defaultValidator, fromExData, reValidateExData, toExData } from '../../../../utils/excel.utils'
+import { defaultValidator, fromExData, reValidateExData, toExData } from '@condo/domains/common/utils/excel.utils'
 import { useIntl } from '@core/next/intl'
 import { useImmer } from 'use-immer'
 import styled from '@emotion/styled'
