+++ apps/condo/domains/common/utils/ecmascript.utils.js
