+++ apps/condo/domains/common/utils/ecmascript.utils.spec.js
