+++ apps/condo/domains/common/utils/firebase.back.utils.js
