+++ apps/condo/domains/common/utils/mediaQuery.utils.spec.js
