+++ apps/condo/domains/common/utils/pdf.ts
