+++ apps/condo/domains/common/utils/url.utils.js
