+++ apps/condo/domains/common/utils/url.utils.spec.js
