+++ apps/condo/domains/common/utils/validation.utils.js
