+++ apps/condo/domains/property/gql.js
