+++ apps/condo/domains/ticket/components/BaseTicketForm/index.tsx
@@ -7,7 +7,7 @@ import React from 'react'
 import { ITicketFormState } from '@condo/domains/ticket/utils/clientSchema/Ticket'
 import { colors } from '../../../common/constants/style'
 import { FormWithAction } from '../../../common/components/containers/FormList'
-import { searchEmployee, searchProperty, searchTicketClassifier } from '../../../../utils/clientSchema/Ticket/search'
+import { searchEmployee, searchProperty, searchTicketClassifier } from '../../utils/clientSchema/search'
 import { FocusContainer } from '@condo/domains/common/components/FocusContainer'
 import { GraphQlSearchInput } from '@condo/domains/common/components/GraphQlSearchInput'
 import { LabelWithInfo } from '@condo/domains/common/components/LabelWithInfo'
