+++ apps/condo/domains/ticket/components/TicketForm/CreateTicketForm.tsx
@@ -7,7 +7,7 @@ import { useRouter } from 'next/router'
 import { useAuth } from '@core/next/auth'
 import { useIntl } from '@core/next/intl'
 import { BaseTicketForm } from '../BaseTicketForm'
-import { Button } from '@condo/domains/components/components/Button'
+import { Button } from '@condo/domains/common/components/Button'
 import { ErrorsContainer } from '../BaseTicketForm/ErrorsContainer'
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore
