+++ apps/condo/domains/ticket/gql.js
@@ -11,9 +11,13 @@ const COMMON_FIELDS = 'id dv sender v deletedAt newId createdBy { id name } upda
 const TICKET_FIELDS = `{ organization { id name } property { id name address } unitName status { id name type organization { id } } statusReopenedCounter statusUpdatedAt statusReason number client { id name } clientName clientEmail clientPhone operator { id name } assignee { id name } executor { id name } watchers { id name } classifier { id name organization { id } parent { id name } } details related { id details } isEmergency isPaid meta source { id name type } sourceMeta ${COMMON_FIELDS} }`
 const Ticket = generateGqlQueries('Ticket', TICKET_FIELDS)
 
+const TICKET_STATUS_FIELDS = '{ id dv name type }'
+const TicketStatus = generateGqlQueries('TicketStatus', TICKET_STATUS_FIELDS)
+
 /* AUTOGENERATE MARKER <CONST> */
 
 module.exports = {
     Ticket,
+    TicketStatus,
 /* AUTOGENERATE MARKER <EXPORTS> */
 }
