+++ apps/condo/domains/ticket/utils/helpers.ts
