+++ apps/condo/domains/user/schema/Auth.js
@@ -1,7 +1,7 @@
 const { getById, getSchemaCtx } = require('@core/keystone/schema')
 const { GQLCustomSchema } = require('@core/keystone/schema')
 
-const { admin } = require('../../../utils/firebase.back.utils')
+const { admin } = require('@condo/domains/common/utils/firebase.back.utils')
 
 const AuthenticateUserWithFirebaseIdTokenService = new GQLCustomSchema('AuthenticateUserWithFirebaseIdTokenService', {
     types: [
