+++ apps/condo/pages/auth/signin.tsx
@@ -11,8 +11,8 @@ import { useIntl } from '@core/next/intl'
 import { useMutation } from '@core/next/apollo'
 
 import { TopMenuOnlyLayout } from '@condo/domains/common/components/containers/BaseLayout'
-import { getQueryParams } from '../../utils/url.utils'
-import { runMutation } from '../../utils/mutations.utils'
+import { getQueryParams } from '@condo/domains/common/utils/url.utils'
+import { runMutation } from '@condo/domains/common/utils/mutations.utils'
 
 import { AuthState, PhoneAuthForm } from './register'
