+++ apps/condo/pages/employee/index.tsx
@@ -7,8 +7,8 @@ import { useIntl } from '@core/next/intl'
 
 import { PageContent, PageHeader, PageWrapper } from '@condo/domains/common/components/containers/BaseLayout'
 import { OrganizationRequired } from '@condo/domains/common/components/containers/OrganizationRequired'
-import { runMutation } from '../../utils/mutations.utils'
-import { emailValidator, nameValidator, phoneValidator } from '../../utils/excel.utils'
+import { runMutation } from '@condo/domains/common/utils/mutations.utils'
+import { emailValidator, nameValidator, phoneValidator } from '@condo/domains/common/utils/excel.utils'
 import { OrganizationEmployee, useInviteNewOrganizationEmployee } from '../../utils/clientSchema/Organization'
 
 import {
