--- apps/condo/utils/clientSchema/Ticket/index.ts
@@ -1,1 +1,0 @@
-export * as TicketStatus from './TicketStatus'
